first attempt at data handling restructure
