+++ src/qicklab/analysis/ssf.py
@@ -2,9 +2,10 @@ import os, datetime
 
 import numpy as np
 
+from ..datahandling.datafile_tools import load_h5_data
 from ..utils.ana_utils  import rotate_and_threshold
 from ..utils.data_utils import process_h5_data
-from ..utils.file_utils import load_from_h5_with_shotdata
+# from ..utils.file_utils import load_from_h5_with_shotdata
 
 class ssf:
     def __init__(self,data_dir, dataset, QubitIndex, folder="study_data", expt_name ="ss_ge"):
@@ -29,7 +30,7 @@ class ssf:
         theta = []
 
         for h5_file in h5_files:
-            load_data = load_from_h5_with_shotdata(os.path.join(data_path, h5_file), 'SS', save_r=1)
+            load_data = load_h5_data(os.path.join(data_path, h5_file), 'SS', save_r=1)
             dates.append(datetime.datetime.fromtimestamp(load_data['SS'][self.QubitIndex].get('Dates', [])[0][0]))
 
             I_g.append(process_h5_data(load_data['SS'][self.QubitIndex].get('I_g', [])[0][0].decode()))
